pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "omni-python-library"
version = "0.1.0"
description = "Omni Python Library"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0.0",
    "python-arango>=7.5.0",
    "redis>=4.5.0",
    "cachetools>=5.3.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "types-redis",
    "types-cachetools",
    "pytest>=7.0.0",
]

[tool.black]
line-length = 120
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["omni_python_library"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.setuptools.packages.find]
where = ["src"]
